First GitHub commit of my javascript collision testing work.
Program revolves around objects derived from constructor functions in gamelibrary.js that areprinted on a canvas. All objects have different types of collision with the player. Map entities collide with the player and prevent further movement in the direction of the object, and one moving "monster" object takes a life from the player object and resets its position.